==> 02 - Normalizacao_dataset_v02.ipynb <==
# Normalização dos dados
-----------------------------------------------------------------------------------------------------------------------------

# Bibliotecas Utilizadas 

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)
import seaborn as sns # visualize
import matplotlib.pyplot as plt
from sklearn.decomposition import PCA
pd.set_option('display.max_columns', None)

# Carga do dataset original

st_port = pd.read_csv('student-por.csv', header = 0)
frames = [st_port]
dataset = pd.concat(frames, sort=False)
print("Length result dataset: {} rows".format(len(dataset)))

dataset.head()

# Cria a coluna is_approved

Esta coluna identifica se o aluno foi ou não aprovado na matéria
Conforme falamos, vamos criar uma nova coluna de aprovação x reprovação. 
Eu considerei a média da escola para definir se o aluno estaria aprovado ou reprovado.

for dataset in frames:
    ave = sum(dataset.G3)/float(len(dataset))
    dataset['is_approved'] = [1 if i > ave else 0 for i in dataset.G3]

dataset.head()

# Remove as colunas G1, G2 e G3
Será utilizada a coluna is_approved

dataset.drop(['G1','G2','G3'], axis=1, inplace=True)

dataset.head()

dataset.columns

# Aplicar one-hot enconding nas colunas
'Medu','Fedu','Mjob','Fjob','reason','guardian'

# converte para binário
one_hot_columns = ['Medu','Fedu','Mjob','Fjob','reason','guardian']

dataset = dataset.copy()
dataset = pd.get_dummies(dataset, columns=['Medu','Fedu','Mjob','Fjob','reason','guardian'], 
                         prefix = ['Medu','Fedu','Mjob','Fjob','reason','guardian'])


dataset.head()

# Converte de nominal para binário

print("school: " , dataset.school.unique())
print("sex: " , dataset.sex.unique())
print("address: " , dataset.address.unique())
print("famsize: " , dataset.famsize.unique())
print("schoolsup: " , dataset.schoolsup.unique())
print("famsup: " , dataset.famsup.unique())
print("paid: " , dataset.paid.unique())
print("activities: " , dataset.activities.unique())
print("nursery: " , dataset.nursery.unique())
print("higher: " , dataset.higher.unique())
print("internet: " , dataset.internet.unique())
print("romantic: " , dataset.romantic.unique())
print("Pstatus: " , dataset.Pstatus.unique())

dataset['school'] = dataset['school'].map( {'GP': 0, 'MS': 1} ).astype(int)
dataset['sex'] = dataset['sex'].map( {'F': 0, 'M': 1} ).astype(int)
dataset['address'] = dataset['address'].map( {'U': 0, 'R': 1} ).astype(int)
dataset['famsize'] = dataset['famsize'].map( {'GT3': 0, 'LE3': 1}).astype(int)
dataset['Pstatus'] = dataset['Pstatus'].map( {'A': 0, 'T': 1}).astype(int)

binary_columns = ['schoolsup', 'famsup', 'paid', 'activities', 'nursery','higher', 'internet', 'romantic']

for c in binary_columns:
    dataset[c] = dataset[c].map( {'yes': 1, 'no': 0} ).astype(int)

dataset.head()

# Aplica min-max nas colunas

dataset.columns

from sklearn import preprocessing

x = dataset.values  #returns a numpy array

min_max_scaler = preprocessing.MinMaxScaler()

x_scaled = min_max_scaler.fit_transform(x.astype(float))
dataset_normal = pd.DataFrame(x_scaled, columns=['school', 'sex', 'age', 'address', 'famsize', 'Pstatus', 'traveltime',
                                   'studytime', 'failures', 'schoolsup', 'famsup', 'paid', 'activities',
                                   'nursery', 'higher', 'internet', 'romantic', 'famrel', 'freetime',
                                   'goout', 'Dalc', 'Walc', 'health', 'absences', 'is_approved', 'Medu_0',
                                   'Medu_1', 'Medu_2', 'Medu_3', 'Medu_4', 'Fedu_0', 'Fedu_1', 'Fedu_2',
                                   'Fedu_3', 'Fedu_4', 'Mjob_at_home', 'Mjob_health', 'Mjob_other',
                                   'Mjob_services', 'Mjob_teacher', 'Fjob_at_home', 'Fjob_health',
                                   'Fjob_other', 'Fjob_services', 'Fjob_teacher', 'reason_course',
                                   'reason_home', 'reason_other', 'reason_reputation', 'guardian_father',
                                   'guardian_mother', 'guardian_other'])

# Temos um novo dataset normalizado: dataset_normal

print("Rows: {}".format(len(dataset)))
print("Columns: {}" .format(len(dataset.columns)))
dataset_normal.head()

# Exporta a base normalizada

dataset_normal.to_csv("dataset-normalizado.csv", index=False)
print("Rows: {}".format(len(dataset)))
print("Columns: {}" .format(len(dataset.columns)))

dataset_normal['is_approved'].value_counts()
