==> 04 Features Selection - Correlation Matrix.ipynb <==
# Seleção de Características - Matrix de Correlação

### Bibliotecas Utilizadas 

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)
import seaborn as sns # visualize
import matplotlib.pyplot as plt

# Carga da base original
dataset = pd.read_csv('dataset-normalizado.csv', header = 0)

# Reordena colocando a coluna is_approved no final da tabela
dataset = dataset[['school', 'sex', 'age', 'address', 'famsize', 'Pstatus',
       'traveltime', 'studytime', 'failures', 'schoolsup', 'famsup', 'paid',
       'activities', 'nursery', 'higher', 'internet', 'romantic', 'famrel',
       'freetime', 'goout', 'Dalc', 'Walc', 'health', 'absences', 'Medu_0',
       'Medu_1', 'Medu_2', 'Medu_3', 'Medu_4', 'Fedu_0', 'Fedu_1', 'Fedu_2',
       'Fedu_3', 'Fedu_4', 'Mjob_at_home', 'Mjob_health', 'Mjob_other',
       'Mjob_services', 'Mjob_teacher', 'Fjob_at_home', 'Fjob_health',
       'Fjob_other', 'Fjob_services', 'Fjob_teacher', 'reason_course',
       'reason_home', 'reason_other', 'reason_reputation', 'guardian_father',
       'guardian_mother', 'guardian_other','is_approved']]

dataset.head()

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

## Criamos um dataset com a matriz de correlação

cor_matrix = dataset.corr()

cor_matrix.head()

is_approved_cor = cor_matrix.iloc[-1:]

is_approved_cor

is_approved_cor = is_approved_cor.transpose()

## Buscamos as 10 características com maior correlação com o *target*

fs_10_cor = is_approved_cor.nlargest(11, 'is_approved')[1:]

fs_10_cor

sorted(fs_10_cor.axes[0].tolist(), key=lambda s: s.lower())

## Criamos um novo dataset apenas com as características selecionadas e a coluna de target

fs_4_dataset = dataset[['activities',
 'Fedu_4',
 'Fjob_teacher',
 'higher',
 'internet',
 'Medu_4',
 'Mjob_health',
 'Mjob_teacher',
 'reason_reputation',
 'studytime',
 'is_approved']]

fs_4_dataset.head()

# plt.figure(figsize=(100, 100))
# sns.heatmap(dataset.corr(),
#             annot = True,
#             fmt = '.2f',
#             cmap='Blues')
# plt.title('Correlação entre variáveis do dataset')
# plt.show()
