==> 05 Embeded - Select from Model.ipynb <==
### Bibliotecas Utilizadas 
import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)
import seaborn as sns # visualize
import matplotlib.pyplot as plt

# Carga da base original
dataset = pd.read_csv('dataset-normalizado.csv', header = 0)

# Reordena colocando a coluna is_approved no final da tabela
dataset = dataset[['school', 'sex', 'age', 'address', 'famsize', 'Pstatus',
       'traveltime', 'studytime', 'failures', 'schoolsup', 'famsup', 'paid',
       'activities', 'nursery', 'higher', 'internet', 'romantic', 'famrel',
       'freetime', 'goout', 'Dalc', 'Walc', 'health', 'absences', 'Medu_0',
       'Medu_1', 'Medu_2', 'Medu_3', 'Medu_4', 'Fedu_0', 'Fedu_1', 'Fedu_2',
       'Fedu_3', 'Fedu_4', 'Mjob_at_home', 'Mjob_health', 'Mjob_other',
       'Mjob_services', 'Mjob_teacher', 'Fjob_at_home', 'Fjob_health',
       'Fjob_other', 'Fjob_services', 'Fjob_teacher', 'reason_course',
       'reason_home', 'reason_other', 'reason_reputation', 'guardian_father',
       'guardian_mother', 'guardian_other', 'is_approved']]

len(dataset.columns)

## Biblioteca para selecao de caracteristicas

from sklearn.feature_selection import SelectFromModel

from sklearn.linear_model import LogisticRegression

X = dataset.drop(['is_approved'], axis=1)
y = dataset.is_approved
feature_name = X.columns.tolist()

from sklearn.feature_selection import SelectFromModel

clf = DecisionTreeClassifier()

### Definimos que o threshold a ser considerado deve ser a mediana e o seletor deve selecionar as 10 melhores características

dt_selector = SelectFromModel(clf, threshold='median', max_features=10)

### Parâmetros utilizados no seletor de características

dt_selector.get_params()

dt_selector.fit(X, y)

dt_support = dt_selector.get_support()

print(dt_support)

dt_threshold = dt_selector.threshold_
print(dt_threshold)

dt_coef = dt_selector.estimator_.feature_importances_
print(dt_coef)

Apenas as características com `feature_importance >= dt_threshold` foram selecionadas

dt_features = X.loc[:,dt_support].columns.tolist()

print(str(len(dt_features)), 'selected features')

dt_features

## Logistic Regression

# lr_selector = SelectFromModel(LogisticRegression(penalty="l2"), '1.25*median')
# lr_selector.fit(X, y)
# embeded_lr_support = embeded_lr_selector.get_support()
# print(embeded_lr_support)

# embeded_lr_threshold = embeded_lr_selector.threshold_
# # print(embeded_lr_threshold)

# embeded_lr_coef = embeded_lr_selector.estimator_.coef_
# print(embeded_lr_coef)

# embeded_lr_feature = X.loc[:,embeded_lr_support].columns.tolist()
# print(str(len(embeded_lr_feature)), 'selected features')

# embeded_lr_feature
