==> 06 - k-fold cross validation + naive Bayes v02.ipynb <==
# Naive Bayes
-------------------------------------------------------------------------------------------------------------

# Bibliotecas Necessárias

import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)
import numpy as np
import seaborn as sns # visualize
import matplotlib.pyplot as plt

from sklearn.model_selection import StratifiedKFold
from sklearn.naive_bayes import GaussianNB
from sklearn import metrics

# Funções Auxiliares

describe_dataset() : realiza o cálculo das proporções de classes do dataset original

def describe_dataset(X, y, k):
    # get dataset rows: instances , columns: features
    rows, columns = X.shape
    # get proportion from target
    (unique, counts) = np.unique(y, return_counts=True) 
    # calculate proportion
    prop_neg = int(counts[0]/rows*100)
    prop_pos = int(counts[1]/rows*100)

    print("k = {}, Dataset: {} positivas, {} negativas ({}% x {}%)".format(k, counts[1], counts[0], prop_pos, prop_neg))

get_classes_from_index() : realiza o cálculo das proporções de classes dos folds criados

def get_classes_from_index(y, skf):
    _, y_idx, y_inv = np.unique(y, return_index=True, return_inverse=True)
    y_counts = np.bincount(y_inv)
    _, class_perm = np.unique(y_idx, return_inverse=True)
    y_encoded = class_perm[y_inv]
    y_order = np.sort(y_encoded)
    n_classes = len(y_idx)
    allocation = np.asarray(
            [np.bincount(y_order[i::skf.n_splits], minlength=n_classes)
             for i in range(skf.n_splits)])

    for idx, f in enumerate(allocation):
        count_neg = int(f[0])
        count_pos = int(f[1])
        total = count_neg+count_pos
        prop_temp_neg = int(count_neg/total*100)
        prop_temp_pos = int(count_pos/total*100)
        print("Fold {}: Pos: {}, Neg: {}, Total: {}, Proporção: {}% x {}%".format(idx, count_pos, count_neg, total, prop_temp_pos, prop_temp_neg))

def stratified_k_fold(X, y, k, reports):
    """
    Parameters
    ----------    
    X : array-like, shape (n_samples, n_features)
        Training data, where n_samples is the number of samples
        and n_features is the number of features.
    y : array-like, of length n_samples
        The target variable for supervised learning problems.
    k : int
        Determines the number of folds.

    """

    skf = StratifiedKFold(n_splits=k)
    describe_dataset(X, y, k)
    get_classes_from_index(y, skf)
    print(skf)
    
    ### create naive bayes classifier
    clf = GaussianNB()
        
    
    ### Executa o treino e teste para k folds
    for train_index, test_index in skf.split(X, y):
        
        print("\nTRAIN: {}  TEST: {}".format(len(train_index), len(test_index)))
        X_train, X_test = X[train_index], X[test_index]
        y_train, y_test = y[train_index], y[test_index]
        
        ### train classifier
        clf.fit(X_train, y_train)
                  
        ### calculate metrics
        y_predicted = clf.predict(X_test)
        report_dict = metrics.classification_report(y_test, y_predicted, output_dict=True)
        report_str = metrics.classification_report(y_test, y_predicted)
        reports.append(report_dict)
        print(report_str)
    
    

# Execução base: Todas as características

reports_dict_all_features = []

df = pd.read_csv('dataset-normalizado.csv', header = 0)
X = df.drop('is_approved', axis=1).to_numpy() # DATASET
y = df['is_approved'].to_numpy() # target
stratified_k_fold(X, y, k=10, reports=reports_dict_all_features)

Gera o resultado do processamento da base considerando todas as características

result_all_features = pd.DataFrame(pd.DataFrame(reports_dict_all_features)['1.0'].to_list())
accuracy_all_features = pd.DataFrame(pd.DataFrame(reports_dict_all_features)['accuracy'])
result_all_features['accuracy'] = accuracy_all_features
result_all_features

result_mean_all_features = pd.DataFrame(result_all_features.mean(), columns=['All Features'])
result_mean_all_features
# result_mean_all_features.transpose()

# Execução Base: PCA

reports_dict_pca = []

df = pd.read_csv('dataset-pca.csv', header = 0)
X = df.drop('is_approved', axis=1).to_numpy() # DATASET
y = df['is_approved'].to_numpy() # target
stratified_k_fold(X, y, k=10, reports=reports_dict_pca)

# Gera a tabela com os indicadores do resultados dos k-fold

result_pca = pd.DataFrame(pd.DataFrame(reports_dict_pca)['1.0'].to_list())
accuracy_pca = pd.DataFrame(pd.DataFrame(reports_dict_pca)['accuracy'])
result_pca['accuracy'] = accuracy_pca
result_pca

result_mean_pca = pd.DataFrame(result_pca.mean(), columns=['PCA'])
result_mean_pca
# result_mean_pca.transpose()

# Execução Base: Chi Squared (K-Best)

reports_dict_chi = []

df = pd.read_csv('dataset-fs-chi-squared.csv', header = 0)
X = df.drop('is_approved', axis=1).to_numpy() # DATASET
y = df['is_approved'].to_numpy() # target
stratified_k_fold(X, y, k=10, reports=reports_dict_chi)

Gera a tabela com os indicadores do resultados dos k-fold

result_chi = pd.DataFrame(pd.DataFrame(reports_dict_chi)['1.0'].to_list())
accuracy_chi = pd.DataFrame(pd.DataFrame(reports_dict_chi)['accuracy'])
result_chi['accuracy'] = accuracy_chi
result_chi

result_mean_chi = pd.DataFrame(result_chi.mean(), columns=['Chi Squared'])
result_mean_chi
# result_mean_pca.transpose()

# Execução Base: recursive-feature

reports_dict_recursive = []

df = pd.read_csv('dataset-fs-recursive-feature.csv', header = 0)
X = df.drop('is_approved', axis=1).to_numpy() # DATASET
y = df['is_approved'].to_numpy() # target
stratified_k_fold(X, y, k=10, reports=reports_dict_recursive)

result_recursive = pd.DataFrame(pd.DataFrame(reports_dict_recursive)['1.0'].to_list())
accuracy_recursive = pd.DataFrame(pd.DataFrame(reports_dict_recursive)['accuracy'])
result_recursive['accuracy'] = accuracy_chi
result_recursive

result_mean_recursive = pd.DataFrame(result_recursive.mean(), columns=['Recursive Feature'])
result_mean_recursive
# result_mean_pca.transpose()

# Junta os resultados gerados de cada base

result = pd.concat([result_mean_all_features, result_mean_pca, result_mean_chi, result_mean_recursive], axis=1)

print("Média das métricas geradas pelo processamento de cada dataset")
result.iloc[[0,1,2,4]]
result.transpose()
