==> 06 Features Selection Comparison.ipynb <==
import pandas as pd
import numpy as np

dataset_chi_squared = pd.read_csv('dataset-fs-chi-squared.csv', header=0)
dataset_chi_squared.name = "MÉTODO 01 - FILTER - CHI SQUARED"
dataset_rfe = pd.read_csv('dataset-fs-rfe.csv', header=0)
dataset_rfe.name = "MÉTODO 02 - WRAPPER - RFE"
dataset_sm = pd.read_csv('dataset-fs-select-model.csv', header=0)
dataset_sm.name = "MÉTODO 03 - EMBEDED - SELECT FROM MODEL"

# Carga da base original
dataset_normal = pd.read_csv('dataset-normalizado.csv', header = 0)

# Reordena colocando a coluna is_approved no final da tabela
dataset_normal = dataset_normal[['school', 'sex', 'age', 'address', 'famsize', 'Pstatus',
       'traveltime', 'studytime', 'failures', 'schoolsup', 'famsup', 'paid',
       'activities', 'nursery', 'higher', 'internet', 'romantic', 'famrel',
       'freetime', 'goout', 'Dalc', 'Walc', 'health', 'absences', 'Medu_0',
       'Medu_1', 'Medu_2', 'Medu_3', 'Medu_4', 'Fedu_0', 'Fedu_1', 'Fedu_2',
       'Fedu_3', 'Fedu_4', 'Mjob_at_home', 'Mjob_health', 'Mjob_other',
       'Mjob_services', 'Mjob_teacher', 'Fjob_at_home', 'Fjob_health',
       'Fjob_other', 'Fjob_services', 'Fjob_teacher', 'reason_course',
       'reason_home', 'reason_other', 'reason_reputation', 'guardian_father',
       'guardian_mother', 'guardian_other']]

column_names = list(dataset_normal.columns.values)

datasets = [dataset_chi_squared, dataset_rfe, dataset_sm]

df_feature_names = pd.DataFrame(column_names, columns =['feature'])
df_feature_names.head()

df1 = pd.read_csv('chi-squared.csv', header=0)
print(df1.shape)

df2 = pd.read_csv('rfe.csv', header=0)
print(df2.shape)

df3 = pd.read_csv('sm.csv', header=0)
print(df3.shape)

concatenated_dataframes = pd.concat([df_feature_names, df1, df2, df3], axis=1)

concatenated_dataframes

delete_list = []
for index, row in concatenated_dataframes.iterrows():
    if row['FILTER - CHI SQUARED'] == False and row ['WRAPPER - RFE'] == False and row['EMBEDED - SELECT FROM MODEL'] == False:
        delete_list.append(index)
    
print(delete_list)

concatenated_dataframes = concatenated_dataframes.drop(concatenated_dataframes.index[delete_list])

concatenated_dataframes

dataset_chi_squared.head()

dataset_rfe.head()

dataset_sm.head()

## 1. SVM `chi-squared` 

from sklearn.model_selection import train_test_split # Import train_test_split function
from sklearn.svm import SVC
from sklearn import metrics
from sklearn import model_selection
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import cross_val_predict

for df in datasets:
    print(df.name)
    X = df.drop('is_approved',axis=1)
    y = df['is_approved']

    # Split dataset into training set and test set
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=1) # 70% training and 30% test

#     print('train size is %i'%y_train.shape[0])
#     print('test size is %i'%y_test.shape[0])
    
    classifier = SVC(random_state=0, kernel='linear',decision_function_shape='ovo')

    classifier.fit(X_train,y_train)

    #Predict the response for test dataset
    y_pred = classifier.predict(X_test)
    
    # Model Accuracy, how often is the classifier correct?
    print("    Accuracy:",metrics.accuracy_score(y_test, y_pred))

    report=metrics.classification_report(y_test,y_pred)

    clf_name=['SVM']

    print('    Reporting for %s:'%clf_name)

    print(report)
